Reorder columns and add width constraints for alerts table
For now the timestamp, source, problem type, object
and parent object columns have a set max width given
by either the elements currently in the table, or by
a fixed size (the timestamp column).

The description column width is not constrained at all
but will use all the available space left for it.

The columns are now in the following order:
Timestamp,Source,Problem Type,Object,Parent Object,Description

Reasoning:
- Timestamp should be to the left because it is fixed size
  and always relevant
- Source and Problem Type should probably be next to each
  other, but I'm not sure about their relative ordering.
- Object and Parent Object should be next to each other,
  not sure about their relative ordering either.

This technically closes #77