Create and use FilteredIncidentProvider
This takes the existing incident context from the
parent(s), and provides an incident context that
wraps this one but with incidents not matching
using the filterMatcher() removed.

Because of how the useIncidents() hook works,
when you do an API call that modifies an incident,
it will update the incident context value with the new
value. When this value doesn't match the current filter,
(which it doesn't check for at all) we now will display
incidents that doesn't match the filter.